@@ -1,6 +1,6 @@
 {
   "name": "@joplin/app-desktop",
-  "version": "3.1.18",
+  "version": "3.2.0",
   "description": "Joplin for Desktop",
   "main": "main.js",
   "private": true,
@@ -125,19 +125,19 @@
   "devDependencies": {
     "7zip-bin": "5.2.0",
     "@electron/rebuild": "3.6.0",
-    "@joplin/default-plugins": "~3.1",
-    "@joplin/tools": "~3.1",
-    "@playwright/test": "1.44.1",
+    "@joplin/default-plugins": "~3.2",
+    "@joplin/tools": "~3.2",
+    "@playwright/test": "1.45.3",
     "@testing-library/react-hooks": "8.0.1",
     "@types/jest": "29.5.12",
-    "@types/node": "18.19.39",
+    "@types/node": "18.19.42",
     "@types/react": "18.3.3",
     "@types/react-dom": "18.3.0",
     "@types/react-redux": "7.1.33",
     "@types/styled-components": "5.1.32",
     "@types/tesseract.js": "2.0.0",
     "axios": "^1.7.7",
-    "electron": "29.4.5",
+    "electron": "32.2.0",
     "electron-builder": "24.13.3",
     "glob": "10.4.5",
     "gulp": "4.0.2",
@@ -155,16 +155,16 @@
     "@electron/remote": "2.1.2",
     "@fortawesome/fontawesome-free": "5.15.4",
     "@joeattardi/emoji-button": "4.6.4",
-    "@joplin/editor": "~3.1",
-    "@joplin/lib": "~3.1",
-    "@joplin/renderer": "~3.1",
-    "@joplin/utils": "~3.1",
+    "@joplin/editor": "~3.2",
+    "@joplin/lib": "~3.2",
+    "@joplin/renderer": "~3.2",
+    "@joplin/utils": "~3.2",
     "@sentry/electron": "4.24.0",
     "@types/mustache": "4.2.5",
     "async-mutex": "0.5.0",
     "codemirror": "5.65.9",
     "color": "3.2.1",
-    "compare-versions": "6.1.0",
+    "compare-versions": "6.1.1",
     "countable": "3.0.1",
     "debounce": "1.2.1",
     "electron-updater": "6.2.1",
