@@ -4,7 +4,7 @@ import { _ } from '@joplin/lib/locale';
 import DialogButtonRow from './DialogButtonRow';
 const { themeStyle } = require('@joplin/lib/theme');
 const Countable = require('@joplin/lib/countable/Countable');
-import markupLanguageUtils from '../utils/markupLanguageUtils';
+import markupLanguageUtils from '@joplin/lib/utils/markupLanguageUtils';
 import Dialog from './Dialog';
 
 interface NoteContentPropertiesDialogProps {
