@@ -14,6 +14,7 @@ import useFocusHandler from './hooks/useFocusHandler';
 import useOnRenderItem from './hooks/useOnRenderItem';
 import { ListItem } from './types';
 import useSidebarCommandHandler from './hooks/useSidebarCommandHandler';
+import useOnRenderListWrapper from './hooks/useOnRenderListWrapper';
 
 interface Props {
 	dispatch: Dispatch;
@@ -39,11 +40,12 @@ const FolderAndTagList: React.FC<Props> = props => {
 		listItems: listItems,
 	});
 
-	const [selectedListElement, setSelectedListElement] = useState<HTMLElement|null>(null);
+	const listContainerRef = useRef<HTMLDivElement|null>(null);
 	const onRenderItem = useOnRenderItem({
 		...props,
 		selectedIndex,
-		onSelectedElementShown: setSelectedListElement,
+		listItems,
+		containerRef: listContainerRef,
 	});
 
 	const onKeyEventHandler = useOnSidebarKeyDownHandler({
@@ -55,13 +57,16 @@ const FolderAndTagList: React.FC<Props> = props => {
 	});
 
 	const itemListRef = useRef<ItemList<ListItem>>();
-	const { focusSidebar } = useFocusHandler({ itemListRef, selectedListElement, selectedIndex, listItems });
+	const { focusSidebar } = useFocusHandler({ itemListRef, selectedIndex, listItems });
 
 	useSidebarCommandHandler({ focusSidebar });
 
 	const [itemListContainer, setItemListContainer] = useState<HTMLDivElement|null>(null);
+	listContainerRef.current = itemListContainer;
 	const listHeight = useElementHeight(itemListContainer);
 	const listStyle = useMemo(() => ({ height: listHeight }), [listHeight]);
+
+	const onRenderContentWrapper = useOnRenderListWrapper({ selectedIndex, onKeyDown: onKeyEventHandler });
 
 	return (
 		<div
@@ -72,9 +77,15 @@ const FolderAndTagList: React.FC<Props> = props => {
 				className='items'
 				ref={itemListRef}
 				style={listStyle}
+
 				items={listItems}
 				itemRenderer={onRenderItem}
-				onKeyDown={onKeyEventHandler}
+				renderContentWrapper={onRenderContentWrapper}
+
+				// The selected item is the only item with tabindex=0. Always render it
+				// to allow the item list to be focused.
+				alwaysRenderSelection={true}
+				selectedIndex={selectedIndex}
 
 				itemHeight={30}
 			/>
