@@ -1,12 +1,11 @@
 import * as React from 'react';
 import { useCallback } from 'react';
-import { ButtonLevel } from '../../Button/Button';
-import { StyledAddButton, StyledHeader, StyledHeaderIcon, StyledHeaderLabel } from '../styles';
+import { StyledHeader, StyledHeaderIcon, StyledHeaderLabel } from '../styles';
 import { HeaderId, HeaderListItem } from '../types';
-import { _ } from '@joplin/lib/locale';
 import bridge from '../../../services/bridge';
 import MenuUtils from '@joplin/lib/services/commands/MenuUtils';
 import CommandService from '@joplin/lib/services/CommandService';
+import ListItemWrapper, { ListItemRef } from './ListItemWrapper';
 
 const Menu = bridge().Menu;
 const MenuItem = bridge().MenuItem;
@@ -14,9 +13,12 @@ const menuUtils = new MenuUtils(CommandService.instance());
 
 
 interface Props {
+	anchorRef: ListItemRef;
 	item: HeaderListItem;
+	isSelected: boolean;
 	onDrop: React.DragEventHandler|null;
-	anchorRef: React.Ref<HTMLElement>;
+	index: number;
+	itemCount: number;
 }
 
 const HeaderItem: React.FC<Props> = props => {
@@ -42,30 +44,25 @@ const HeaderItem: React.FC<Props> = props => {
 		}
 	}, [itemId]);
 
-	const addButton = <StyledAddButton
-		iconLabel={_('New')}
-		onClick={item.onPlusButtonClick}
-		iconName='fas fa-plus'
-		level={ButtonLevel.SidebarSecondary}
-	/>;
-
 	return (
-		<div
+		<ListItemWrapper
+			containerRef={props.anchorRef}
+			selected={props.isSelected}
+			itemIndex={props.index}
+			itemCount={props.itemCount}
+			expanded={props.item.expanded}
+			onContextMenu={onContextMenu}
+			depth={0}
+			highlightOnHover={false}
 			className='sidebar-header-container'
 			{...item.extraProps}
 			onDrop={props.onDrop}
 		>
-			<StyledHeader
-				onContextMenu={onContextMenu}
-				onClick={onClick}
-				tabIndex={0}
-				ref={props.anchorRef}
-			>
-				<StyledHeaderIcon aria-label='' className={item.iconName}/>
+			<StyledHeader onClick={onClick}>
+				<StyledHeaderIcon aria-label='' role='img' className={item.iconName}/>
 				<StyledHeaderLabel>{item.label}</StyledHeaderLabel>
 			</StyledHeader>
-			{ item.onPlusButtonClick && addButton }
-		</div>
+		</ListItemWrapper>
 	);
 };
 
