@@ -12,6 +12,7 @@ import { ResourceEntity } from './services/database/types';
 import { TextItem } from 'pdfjs-dist/types/src/display/api';
 import replaceUnsupportedCharacters from './utils/replaceUnsupportedCharacters';
 import { FetchBlobOptions } from './types';
+import crypto from './services/e2ee/crypto';
 
 import FileApiDriverLocal from './file-api-driver-local';
 import * as mimeUtils from './mime-utils';
@@ -135,6 +136,7 @@ function shimInit(options: ShimInitOptions = null) {
 	shim.Geolocation = GeolocationNode;
 	shim.FormData = require('form-data');
 	shim.sjclModule = require('./vendor/sjcl.js');
+	shim.crypto = crypto;
 	shim.electronBridge_ = options.electronBridge;
 
 	shim.fsDriver = () => {
@@ -414,11 +416,10 @@ function shimInit(options: ShimInitOptions = null) {
 		return newBody.join('\n\n');
 	};
 
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any -- Old code before rule was applied
-	shim.attachFileToNote = async function(note, filePath, position: number = null, options: any = null) {
+	shim.attachFileToNote = async function(note, filePath, options = {}) {
 		if (!options) options = {};
 		if (note.markup_language) options.markupLanguage = note.markup_language;
-		const newBody = await shim.attachFileToNoteBody(note.body, filePath, position, options);
+		const newBody = await shim.attachFileToNoteBody(note.body, filePath, options.position ?? 0, options);
 		if (!newBody) return null;
 
 		const newNote = { ...note, body: newBody };
