@@ -8,6 +8,7 @@ import { userIdFromUserContentUrl } from '../utils/routeUtils';
 import { getCanShareFolder } from './utils/user';
 import { isUniqueConstraintError } from '../db';
 import Logger from '@joplin/utils/Logger';
+import { PerformanceTimer } from '../utils/time';
 
 const logger = Logger.create('ShareModel');
 
@@ -184,6 +185,7 @@ export default class ShareModel extends BaseModel<Share> {
 	}
 
 	public async updateSharedItems3() {
+		const perfTimer = new PerformanceTimer(logger, 'updateSharedItems3');
 
 		const addUserItem = async (shareUserId: Uuid, itemId: Uuid) => {
 			try {
@@ -214,11 +216,15 @@ export default class ShareModel extends BaseModel<Share> {
 				return;
 			}
 
+			perfTimer.push('handleCreated');
+
 			const shareUserIds = await this.allShareUserIds(share);
 			for (const shareUserId of shareUserIds) {
 				if (shareUserId === change.user_id) continue;
 				await addUserItem(shareUserId, item.id);
 			}
+
+			perfTimer.pop();
 		};
 
 		const handleUpdated = async (change: Change, item: Item, share: Share) => {
@@ -228,30 +234,36 @@ export default class ShareModel extends BaseModel<Share> {
 
 			if (previousShareId === shareId) return;
 
-			const previousShare = previousShareId ? await this.models().share().load(previousShareId) : null;
+			perfTimer.push('handleUpdated');
 
-			if (previousShare) {
-				const shareUserIds = await this.allShareUserIds(previousShare);
-				for (const shareUserId of shareUserIds) {
-					if (shareUserId === change.user_id) continue;
-					try {
-						await removeUserItem(shareUserId, item.id);
-					} catch (error) {
-						if (error.httpCode === ErrorNotFound.httpCode) {
-							logger.warn('Could not remove a user item because it has already been removed:', error);
-						} else {
-							throw error;
+			try {
+				const previousShare = previousShareId ? await this.models().share().load(previousShareId) : null;
+
+				if (previousShare) {
+					const shareUserIds = await this.allShareUserIds(previousShare);
+					for (const shareUserId of shareUserIds) {
+						if (shareUserId === change.user_id) continue;
+						try {
+							await removeUserItem(shareUserId, item.id);
+						} catch (error) {
+							if (error.httpCode === ErrorNotFound.httpCode) {
+								logger.warn('Could not remove a user item because it has already been removed:', error);
+							} else {
+								throw error;
+							}
 						}
 					}
 				}
-			}
 
-			if (share) {
-				const shareUserIds = await this.allShareUserIds(share);
-				for (const shareUserId of shareUserIds) {
-					if (shareUserId === change.user_id) continue;
-					await addUserItem(shareUserId, item.id);
+				if (share) {
+					const shareUserIds = await this.allShareUserIds(share);
+					for (const shareUserId of shareUserIds) {
+						if (shareUserId === change.user_id) continue;
+						await addUserItem(shareUserId, item.id);
+					}
 				}
+			} finally {
+				perfTimer.pop();
 			}
 		};
 
@@ -265,6 +277,8 @@ export default class ShareModel extends BaseModel<Share> {
 		// that have recently changed, and the performed SQL queries are
 		// index-based.
 		const checkForMissingUserItems = async (shares: Share[]) => {
+			perfTimer.push(`checkForMissingUserItems: ${shares.length} shares`);
+
 			for (const share of shares) {
 				const realShareItemCount = await this.itemCountByShareId(share.id);
 				const shareItemCountPerUser = await this.itemCountByShareIdPerUser(share.id);
@@ -279,6 +293,8 @@ export default class ShareModel extends BaseModel<Share> {
 					}
 				}
 			}
+
+			perfTimer.pop();
 		};
 
 		// This loop essentially applies the change made by one user to all the
@@ -296,20 +312,36 @@ export default class ShareModel extends BaseModel<Share> {
 		// This is probably safer in terms of avoiding race conditions and
 		// possibly faster.
 
-		while (true) {
-			const latestProcessedChange = await this.models().keyValue().value<string>('ShareService::latestProcessedChange');
+		perfTimer.push('Main');
 
+		while (true) {
+			perfTimer.push('Get latestProcessedChange');
+			const latestProcessedChange = await this.models().keyValue().value<string>('ShareService::latestProcessedChange');
+			perfTimer.pop();
+
+			perfTimer.push('Get paginated changes');
 			const paginatedChanges = await this.models().change().allFromId(latestProcessedChange || '');
+			perfTimer.pop();
 			const changes = paginatedChanges.items;
 
 			if (!changes.length) {
+				perfTimer.push('Set latestProcessedChange');
 				await this.models().keyValue().setValue('ShareService::latestProcessedChange', paginatedChanges.cursor);
+				perfTimer.pop();
 			} else {
+				perfTimer.push(`Load items for ${changes.length} changes`);
 				const items = await this.models().item().loadByIds(changes.map(c => c.item_id));
+				perfTimer.pop();
 				const shareIds = unique(items.filter(i => !!i.jop_share_id).map(i => i.jop_share_id));
-				const shares = await this.models().share().loadByIds(shareIds);
 
+				perfTimer.push(`Load ${shareIds.length} shares`);
+				const shares = await this.models().share().loadByIds(shareIds);
+				perfTimer.pop();
+
+				perfTimer.push('Change processing transaction');
 				await this.withTransaction(async () => {
+					perfTimer.push(`Processing ${changes.length} changes`);
+
 					for (const change of changes) {
 						const item = items.find(i => i.id === change.item_id);
 
@@ -335,11 +367,16 @@ export default class ShareModel extends BaseModel<Share> {
 					await checkForMissingUserItems(shares);
 
 					await this.models().keyValue().setValue('ShareService::latestProcessedChange', paginatedChanges.cursor);
+
+					perfTimer.pop();
 				}, 'ShareService::updateSharedItems3');
+				perfTimer.pop();
 			}
 
 			if (!paginatedChanges.has_more) break;
 		}
+
+		perfTimer.pop();
 	}
 
 	public async updateResourceShareStatus(doShare: boolean, _shareId: Uuid, changerUserId: Uuid, toUserId: Uuid, resourceIds: string[]) {
